[Ollie & Luke] Move de-duplication logic to dropzones file
